requirements.txt
discord.py==0.16.12
aiopg==0.14.0
uvloop==0.11.0
aiohttp==1.0.5
websockets==3.4
matplotlib==3.0.1
